move MesosUtilTest to own file
